+++ CHANGELOG.md
@@ -3,4 +3,10 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Fix memory leaks
+- Improve nax-ccuilib compatibility
+
 ## [1.0.0] 2025-10-04